refactor(decisions): extract useAdvancePhase from DecisionProcessStepper
Move the advance-to-next-phase flow (transitionFromPhase mutation, confirm
Modal/Sheet, PostHog tracking, success toast + refresh) into a shared
useAdvancePhase hook so the upcoming overview phase timeline can reuse it.
DecisionProcessStepper now renders the returned confirm dialog and calls
requestAdvance from the stepper's transition affordance. No behavior change.
